gilded-rose: starter with characterization test
