+++ gilded-rose/App/App.csproj
