+++ gilded-rose/App/GildedRose.cs
@@ -1,10 +1,12 @@
 using System.Collections.Generic;
 
-namespace GildedRose.App
+namespace App
 {
+    // Keep this for non regression
     public class GildedRose
     {
         IList<Item> Items;
+
         public GildedRose(IList<Item> Items)
         {
             this.Items = Items;
